// orientation-post-baccalauriat-FrontEnd/src/app/core/enums/institution-type.ts
export enum InstitutionType {
  PUBLIC_UNIVERSITY = 'PUBLIC_UNIVERSITY',               // Université Publique
  PRIVATE_UNIVERSITY = 'PRIVATE_UNIVERSITY',             // Université Privée
  HIGHER_INSTITUTE = 'HIGHER_INSTITUTE',                 // Institut Supérieur
  VOCATIONAL_TRAINING_CENTER = 'VOCATIONAL_TRAINING_CENTER', // Centre de Formation Professionnelle
  ROYAL_ACADEMY = 'ROYAL_ACADEMY',                       // Académie Royale
  EXECUTIVE_TRAINING_INSTITUTION = 'EXECUTIVE_TRAINING_INSTITUTION', // Institution de Formation pour Cadres
  POLYTECHNIC_SCHOOL = 'POLYTECHNIC_SCHOOL',             // École Polytechnique
  RESEARCH_INSTITUTE = 'RESEARCH_INSTITUTE'              // Institut de Recherche
}